Extract RelativeTimeFormat into helpers & use for workspaces
