+++ src/providers/tfc/runProvider.ts
@@ -19,7 +19,7 @@ import {
   TRIGGER_REASON,
 } from '../../terraformCloud/run';
 import { WorkspaceTreeItem } from './workspaceProvider';
-import { GetRunStatusIcon } from './helpers';
+import { GetRunStatusIcon, RelativeTimeFormat } from './helpers';
 import { z } from 'zod';
 
 export class RunTreeDataProvider implements vscode.TreeDataProvider<vscode.TreeItem>, vscode.Disposable {
@@ -196,7 +196,7 @@ async function runMarkdown(item: RunTreeItem) {
   // TODO(fix): the date does not get parsed as Date via zod schema
   const date = z.coerce.date().parse(item.attributes['created-at']);
 
-  const createdAtTime = relativeTimeFormat(date);
+  const createdAtTime = RelativeTimeFormat(date);
 
   if (item.createdBy) {
     markdown.appendMarkdown(`<img src="${item.createdBy?.['avatar-url']}" width="20"> **${item.createdBy?.username}**`);
@@ -242,44 +242,4 @@ _____
 | **Execution Mode** | ${item.workspace.attributes['execution-mode']} |
 `);
   return markdown;
-}
-
-function relativeTimeFormat(d: Date): string {
-  const SECONDS_IN_MINUTE = 60;
-  const SECONDS_IN_HOUR = SECONDS_IN_MINUTE * 60;
-  const SECONDS_IN_DAY = SECONDS_IN_HOUR * 24;
-  const SECONDS_IN_WEEK = SECONDS_IN_DAY * 7;
-  const SECONDS_IN_MONTH = SECONDS_IN_DAY * 30;
-  const SECONDS_IN_YEAR = SECONDS_IN_DAY * 365;
-
-  const rtf = new Intl.RelativeTimeFormat('en', { style: 'long', numeric: 'auto' });
-  const nowSeconds = Date.now() / 1000;
-  const seconds = d.getTime() / 1000;
-  const diffSeconds = nowSeconds - seconds;
-
-  if (diffSeconds < SECONDS_IN_MINUTE) {
-    return rtf.format(-diffSeconds, 'second');
-  }
-  if (diffSeconds < SECONDS_IN_HOUR) {
-    const minutes = Math.round(diffSeconds / SECONDS_IN_MINUTE);
-    return rtf.format(-minutes, 'minute');
-  }
-  if (diffSeconds < SECONDS_IN_DAY) {
-    const hours = Math.round(diffSeconds / SECONDS_IN_HOUR);
-    return rtf.format(-hours, 'hour');
-  }
-  if (diffSeconds < SECONDS_IN_WEEK) {
-    const days = Math.round(diffSeconds / SECONDS_IN_DAY);
-    return rtf.format(-days, 'day');
-  }
-  if (diffSeconds < SECONDS_IN_MONTH) {
-    const weeks = Math.round(diffSeconds / SECONDS_IN_WEEK);
-    return rtf.format(-weeks, 'week');
-  }
-  if (diffSeconds < SECONDS_IN_YEAR) {
-    const months = Math.round(diffSeconds / SECONDS_IN_MONTH);
-    return rtf.format(-months, 'month');
-  }
-  const years = diffSeconds / SECONDS_IN_YEAR;
-  return rtf.format(-years, 'year');
 }
